Fix HttpStreamFactoryTest.Preconnect* when HttpStreamPool is used
For the non-HEv3 code path, these tests use a fake ClientSocketPool
to examine whether ClientSocketPool::RequestSockets() is
called or not, and when called, tests check whether parameters matched
with expected ones. The fake ClientSocketPool's RequestSockets()
immediately completes with OK.

For the HEv3 code path, we use and modify HttpStreamPool::Observer
for that purpose. OnPreconnect() callback is called when preconnect is
requested. If OnPreconnect() returns a value, the preconnect
completes with the value. Since the class becomes something that is
not just a observer, this CL renames the class from Observer to
Delegate.

HttpStreamFactoryTest.PreconnectUnsafePort is still failing. A
subsequent CL will fix it.

Bug: 346835898
Change-Id: I6d4a5c89a701cd3e280ab11f1a36e5aca8409653
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5832863
Reviewed-by: Nidhi Jaju <nidhijaju@chromium.org>
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1351763}