Add a RunDeElevatedNoWait variant that takes a path and parameters
This is a follow-on CL to r1351771.

The new variant allows specifying `path` and `parameters` separately.
`path` specifies the file or object on which to execute the default verb
(typically "open").

This is useful when a `url` is specified for the `path`, in which case
`RunDeElevatedNoWait` would launch the default browser de-elevated.

Low-Coverage-Reason: COVERAGE_UNDERREPORTED UAC bots needed
Change-Id: I3e225590186b33a936cfc76ec164cb0dd5674282
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5839430
Auto-Submit: S Ganesh <ganesh@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Commit-Queue: Greg Thompson <grt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1351893}