[layout-unit] Add specialized `operator*` for `std::integral`
The multiplication of `FixedPoint` and `FixedPoint` requires
dividing the multiplication of raw values by the denominator.
This division is unnecessary when multiplying integral.

This patch adds a specialized `operator*` for `std::integral`
to avoid it. It not only optimize the operation, but also
support `FixedPoint` with `int64_t` storage to multiply by
`std::integral` without needing the `int128_t`.

This patch should have no behavior changes.

Bug: 351092748
Change-Id: I193f255e14cadfd0ec4908ec26a960b08d46045b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5839583
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1351927}