Add parameter for PopupRowWithButtonView to define select behavior
For Autocomplete entries, selecting the "delete" button (by hovering it
or via the keyboard) should unselect the suggestion associated with
the PopupRow.

For plus addresses, that is not the behavior we want. Instead, the
suggestion should remain selected and the suggested plus address
should remain previewed.

This CL adds a parameter to PopupRowWithButtonView that allows
customizing this behavior and uses it for the plus address inline
creation suggestion.

Bug: 362445807
Change-Id: Ia685c69e81898b16049d70b309e5f80215c89dc1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5836735
Commit-Queue: Jan Keitel <jkeitel@google.com>
Reviewed-by: Bruno Braga <brunobraga@google.com>
Cr-Commit-Position: refs/heads/main@{#1351930}