[sync] Move SHARING_MESSAGE controller creation to shared code
For a feature that is implemented under components/, it is best to also
implement the controller-creation logic in components/browser_sync.
Embedders may still inject a null service to explicitly indicate that
the feature shouldn't be enabled.

Change-Id: I4d3459c41af4ca9d8010da0bfe71fd1b96bfa5db
Bug: 344893270
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5839726
Reviewed-by: Rushan Suleymanov <rushans@google.com>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1351946}