iwa: Add an option to install from update manifest on web-app-internals
The installed IWA will be stored in the user profile and will be
displayed accordingly on web-app-internals; this path will be replaced
with the update manifest url (which makes more sense) in one of the
follow-ups -- the update logic will follow.

https://screenshot.googleplex.com/85wpfByTgzamNW3
https://screenshot.googleplex.com/VjACWR8X4HXyFNc

go/release-channel-pinning-for-iwa for more context.

Bug: 358606568

Change-Id: Iaeb932322e878b27edb76e29d26d07f0d399787c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5783531
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Reviewed-by: Robbie McElrath <rmcelrath@chromium.org>
Reviewed-by: Dibyajyoti Pal <dibyapal@chromium.org>
Commit-Queue: Andrew Rayskiy <greengrape@google.com>
Cr-Commit-Position: refs/heads/main@{#1351947}