[sync] Move SEARCH_ENGINES controller creation to shared code
Although there is exactly one embedder that actually uses it, as a
general practice, for a feature that is implemented under components/,
it is best to  implement the controller-creation logic in
components/browser_sync.

This centralizes controller creation and makes it very explicit that
certain embedders do not support the feature (rather than relying on
absence of code).

Change-Id: I2cd6b6a2d1562be704d7a137efe00bf12a09c94e
Bug: 344893270
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5837233
Auto-Submit: Mikel Astiz <mastiz@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Rushan Suleymanov <rushans@google.com>
Cr-Commit-Position: refs/heads/main@{#1351965}