Display IWA version and description in DevTools Apps page
This CL ensures that the version and description of Isolated Web Apps are displayed in the DevTools Apps page (chrome://inspect/#apps).
As IWA versioning is much more rigid than that of other web apps, it may be beneficial to include the version in name in inspect tool.

Bug: 362675374
Change-Id: I13bd2ecce0db876308197273a4dcaece510f121d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5822430
Reviewed-by: Phillis Tang <phillis@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Robert Ferens <rferens@google.com>
Cr-Commit-Position: refs/heads/main@{#1351968}