LinkPreview & fenced frames: Disable LinkPreview context menu option in fenced frame with untrusted network access cutoff
Follow-up CL for crrev.com/c/5823850.

This CL disables the "Preview Link" context menu options if the menu is
opened upon on a link inside a fenced frame whose untrusted network
access is cutoff.

Minor change: rename IsSaveAsItemAllowedByUntrustedNetworkStatus
to IsAllowedByUntrustedNetworkStatus as it is not only used for
save as items.

Bug: 359628054
Change-Id: Ifc974541854b51b70bd8d079898be5724c6e8086
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5835921
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Xiaochen Zhou <xiaochenzh@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1351979}