Avoid layout jumps when transitioning to suggestion loading state
Setting Suggestion::is_loading to true causes (among other things)
the suggestion's icon to be replaced by a throbber.

Prior to plus addresses, loading states were primarily used by
password manager. For password manager, the width of the throbber
and that of the favicon coincided: Both were 16 px.

For a general suggestion, that need not be the case: For example,
both address and plus address suggestions have icons of width 20px.
As a result, transitioning from is_loading = true to
is_loading = false and vice versa caused jumps in layout.

This CL adds a margin to the throbber if the icon is wider than the
throbber. This ensures that the overall width of the suggestion is
the same regardless of the value of Suggestion::is_loading.

Note: For icons that are smaller than 16 px, this will not prevent
layout changes. However, as of now, those are not used.

Before CL: http://screencast/cast/NTY1Njk1MzM0NTgwMjI0MHw1ZjRjNmNlNC1hMA
After CL: http://screencast/cast/NDg0Mjk0OTUwMTI1NTY4MHwxZmFhMjUxYi0zYg

Bug: 362445807
Change-Id: I6aa055990128cef023f68ebb6fc911ed8ef073f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5841670
Reviewed-by: Bruno Braga <brunobraga@google.com>
Commit-Queue: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1351986}