Move the definition of ProfileValueSource
Before this CL, the ProfileValueSource was nested in the AutofillField.
However, since forward declarations of nested types are not possible,
this pattern creates a circular dependency when the ProfileValueSource
is used in the AutofillProfile.

As a solution, the definition is moved to a separate file.

Change-Id: I4714bd4aaaf78602b7d58dc954957e0e2d9b53ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5839686
Reviewed-by: Wolfgang Billenstein <bwolfgang@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Matthias Körber <koerber@google.com>
Cr-Commit-Position: refs/heads/main@{#1352002}