Upload the public key to the DM Server in the main thread.
This is to fix the crash that was caused by crrev/c/5730416 (b/359971492). The crashing line was behind an experiment that was rolled back.

I did an exhaustive manual test and it's now fixed.

1. We need to upload to the dm server on the main thread, as DM token storage and DM service must be called on the main thread.

Therefore, we first create and store the key on a worker thread, then on the original thread, upload the key, then on another worker thread, we cleanup and store the right key. This way, we are "recreating" the persistence delegate, but that's fine, as it's technically a stateless object (could be made into a static class for Mac, but then inheritance and thus sharing code with other platforms would be difficult to implement.)

2. In the browser delegate, we don't need to own the DM token storage. The storage is already set when initializing the platform, and we can retrieve the client id and the DM token directly.

3. Call the right Mac constructor in the factory. Ouch!

Note for e2e test for creating keys (rather than loading):
Delete the key from chrome://connectors-internal, then remove the device from Admin Console, and remove the enrollment token, and then rerun chrome.


Bug: 360848564, 317367921, 359971492
Change-Id: Ib91f1d156934ad3f90f2b02c590f5d29e0869b10
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5800158
Reviewed-by: Sébastien Lalancette <seblalancette@chromium.org>
Commit-Queue: Sébastien Lalancette <seblalancette@chromium.org>
Auto-Submit: Parastoo Geranmayeh <parastoog@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1352011}