svg: optimize SVGPathByteStream construction
This does the following:
1. Changes SVGPathByteStreamBuilder to use a Vector with on-stack
   storage during construction. Doing this means that if the stream
   is small enough (<1k) then only one heap allocation is done (for
   the copy).
2. Changes SVGPathByteStream members to be values, rather than held
   via a unique_ptr. The indirection isn't really buying anything.
3. Changes StylePath to hold the Path via an optional, so that an
   allocation isn't necessary if the Path is allocated.

Doing 1 resulted in a handful of changes to how
SVGPathByteStreamBuilder is used.

It's actually possible to get rid of SVGPathByteStream and use
Vector directly, but surprisingly that was a negative. I suspect
it's because doing that means the return value from
SVGPathByteStreamBuilder::CopyByteStream makes it harder for the
compiler to use RVO (because I wouldn't be constructing the Vector
on the return line). It's possible with a pgo this wouldn't matter,
but I only tried LTO.

This improves construction, but doesn't significantly impact
Speedometer. I did this as part of trying to understand why
the patch for 336154570 hurts Speedometer. I still don't have
an answer to that.

Change-Id: I03db0c83ff6dabd7de4dd40771e14702d6991c57
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5838408
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1352021}