[Tab Switcher] Deduplicate color work
A bunch of color related logic was happening 2-3 times. This isn't
necessary as IS_INCOGNITO is a readable key and IS_SELECTED is always
set to a value along with TAB_ACTION_STATE. Additionally common
properties + selectable or closable properties are always bound so
doing the same logic for the same PropertyKey in both common and one
of the other methods is redundant.

Change-Id: I0f4ac032cf6b0b30ac3b0b480e1ce4d130afe39e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5835994
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Sky Malice <skym@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1352051}