DeferRendererTasksAfterInput: Don't change RAILMode to kResponse
Setting the RAILMode to kResponse for UseCase::kDiscreteInputResponse is
causing regressions in a v8 browsing perf benchmark because the RAILMode
affects GC, specifically to/from loading. While this new UseCase maps
exactly to the R in RAIL, we'd like to separate the deferral aspect of
the experimental UseCase from the side effects of setting the RAILMode.
So, this CL changes the RAILMode mapping for kDiscreteInputResponse to
return whatever the current value is.

As a follow-up, once we know if we can ship this or not, we should
look at either removing RAILMode, since all consumers only care
about going to/from loading, or changing consumers to care about
kResponse (e.g. maybe GC policy should change for discrete input to
be less aggressive, since INP is a short window). Also, the loading
RAILMode ends when (continuous) input is detected -- maybe we can
come up with a better fallback since that affects GC, which can
affect loading performance.

Bug: 350540984, 360151054
Change-Id: I2dc29e25c22ee812aefaa43b34ba9b0b5e5acafe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5828299
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1352116}