[WebAuthn] Add UI for passwords in ambient bubble
The second CL for adding passwords to the ambient bubble, this has the
AmbientSigninController store the password forms, and supplies them to
the view where they are appended in rows following the passkeys.

This is prototype code and currently does not have any tests.

Bug: 358119268
Change-Id: I726d347818dd9ae56c51e4d82d3004ada4a9c4d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5839560
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Adem Derinel <derinel@google.com>
Cr-Commit-Position: refs/heads/main@{#1352120}