[WebAuthn] Don't pass clientDataHash in APP mode on Android
When WebAuthn is being used from a WebView, it isn't able to generate
its own clientData, because this has to be done from the WebAuthn
provider. This is a problem because it attempts to use the clientData
anyway, which won't match the hash that gets signed over from in the
request response.

This corrects that problem.

Bug: 364845737
Change-Id: I52f1c95b0fc224566e41dc8e5722202d89e1157c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5841287
Reviewed-by: Adem Derinel <derinel@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1352132}