[Signin][Android] Fix threading issue when shutting down FakeAccountManagerFacade
Prior to this CL, FakeAccountManagerFacade was always using UI thread
to access mAccountHolders. It meant that getAccessToken had to use
runOnUiThreadBlocking to find the account and fetch the token.
Unfortunately, this was causing issues during the shutdown (see
https://crbug.com/329426609#comment3 for details).
This CL changes the approach used in FakeAccountManagerFacade and
FakeAccountManagerDelegate - now they both use synchronized
collections for storing the list of accounts. This avoids the need to
jump threads in getAccessToken and fixes the issue that was occurring
in sync integration tests. Those tests will be re-enabled in a
subsequent CL.

Bug: 329426609
Change-Id: Id446992bd50cd602f9917df4327aa280e3139ac7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5831008
Auto-Submit: Boris Sazonov <bsazonov@chromium.org>
Reviewed-by: Samar Chehade-Lepleux <samarchehade@google.com>
Commit-Queue: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1352169}