Moved param initialization into finalize() in the Canvas WPT generator
Before this CL, the variant grid's params was initialized the first
time the grid's `params` property was used. This implicit
initialization was confusing because it meant we had to first use
`params` at the exact right time. Using it too soon was a bug (before
the variant's `expected_img` param is configured for instance).

This CL moves the grid's param creation into the `finalize()` method,
removing this confusion.

Bug: 364549423
Change-Id: I00bee16fe9f2d8ab13f5b05b49b511ffa642e746
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5836568
Reviewed-by: Andres Ricardo Perez <andresrperez@chromium.org>
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1352203}