Speculatively avoid a NOTREACHED() crash in PrintBackendWin
In PrintBackendWin::GetPrinterSemanticCapsAndDefaults(), sometimes
the DEVMODE contains an invalid duplex mode value. Although this should
not happen, it does happen in some rare cases. Change the code that
handles this, which is currently a NOTREACHED(), to just ignore the bad
value to prevent a crash. As a result, PrintBackendWin will report the
default duplex mode as "unknown". The code that queries supported duplex
modes remains the same.

Bug: 359253687
Change-Id: Ibe660d6f0390f6258e45171b5d9618298aaaf686
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5793872
Commit-Queue: Lei Zhang <thestig@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Alan Screen <awscreen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1352205}