FLEDGE: WPT test getInterestGroupAdAuctionData more thoroughly.
Unlike previous tests that just checked for emptiness/non-emptiness,
this fully decrypts/decodes/decompresses the message, and checks for
presence of required information.

For now, this relies on the VirtualTestSuite configuring a known
keyset; we will also need to add some sort of test-only API to do so
programmatically to do make it work without one.

Bug: 361027509

Change-Id: I81f7d221c4e85a2ac0263925ea5d7c71455b4e04
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5793398
Reviewed-by: Russ Hamilton <behamilton@google.com>
Reviewed-by: Weizhong Xia <weizhong@google.com>
Commit-Queue: Maks Orlovich <morlovich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1352219}