Move extensions code into WindowController.
I recently added the ExtensionBrowserWindow abstract class to represent
an abstract browser window. But it turns out this already exists (with
many of the same functions) in the existing extensions::WindowController
class.

Having two classes with the same meaning is clearly wrong so this
standardizes on the existing WindowController implementation.

The new BrowserExtensionWindowController code was moved unmodified from the old implementation in ExtensionBrowserWindowDesktop. There should be no behavior change.

Bug: b/361838438
Change-Id: Ice9f5b29fc33601e9168956f1d5ef29189bcc840
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5838054
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Brett Wilson <brettw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1352251}