[TabGroupSync] Remove Test From BookmarkBarViewUnitTest
Removes BookmarkBarViewTest.OnSavedTabGroupUpdateBookmarkBarCallsLayout
because the test is essentially an observer test which we already have
many for this use case. Also it was testing the wrong button.

With the TabGroupsSaveUiUpdate (pinning groups), group 3 is added to the
front of the list. Before that change, groups were added to the end of
the list and is why we were validating button 3 since it will slide over
when button 2 is removed.

Ex:
Old: 1 2 3 -> 1 3
New: 3 2 1 -> 3 1

Change-Id: I8966a1602bf54a6d7475c24195c2f8d941f88dfa
Bug: 350514491
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5838440
Commit-Queue: Darryl James <dljames@chromium.org>
Reviewed-by: Taylor Bergquist <tbergquist@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1352256}