soda: Add tests and placeholders for 4 extra languages.
This adds tests and placeholders in kLanguageComponentConfigs[] . The
array in question is used in so many spots as the basis for decisions
that adding workaround for languages not in the list felt more
challenging than adding to the list when compiling in chromeos.

The list is only modified for chromeos and allows for live caption in
locales that Chrome may not support but ChromeOS with conch
does. Verified and link to video in bug.

Bug: b:364732616
Change-Id: Ib87e70da5427b89327475ba390434fc63b229867
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5836321
Reviewed-by: Evan Liu <evliu@google.com>
Auto-Submit: Rob Schonberger <robsc@chromium.org>
Commit-Queue: Rob Schonberger <robsc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1352278}