Make product specs horizontal carousel arrow buttons always visible
This patch updates the product specifications horizontal carousel to
fit the height of the window, so that the carousel arrows are visible
as the table is scrolled vertically. It also updates the gradient shown behind the new column selector to match the table background color.

Carousel behavior:
http://go/scrcast/NDU0NTc1ODk3MDc3MzUwNHw3ZDVkNmE1Zi1lOQ
Gradient color:
http://screen/Abxi25hWNj3BJHh

Bug: 360181841, 364968218
Change-Id: I797f417ca61e8cc6acc0443a134b5ca4ba85478e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5827748
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Paul Adedeji <pauladedeji@google.com>
Cr-Commit-Position: refs/heads/main@{#1352282}