[tracing] Do not always allocate a TraceValue when parsing scripts
crrev.com/0b4cb631 added some extra args and move the construction of
the traced value out into common code. This causes it to always be
allocated even if tracing is off. Adjust the code so it doesn't actually
allocate when tracing is off.

Change-Id: I4e50c33405ed9952fc49115b31986cf31597b9ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5842015
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1352307}