[CVV Storage Enhancement] Rename histogram for StandaloneCvc form.
Previously when FormType equals to kStandaloneCvcForm, the histogram
variant was "VirtualCard.StandaloneCvc". In this CL, we change it to
generic "StandaloneCvc" so that we can log into this UMA for all
standalone CVC forms (separate logging from the normal credit card
form).

I don't update any current logging logic, just update the UMA name and
add thorough tests to verify the current logging.

There are three kinds of forms:
1) Normal credit card forms:
    FormEvents logged to "Autofill.FormEvents.CreditCard"
2) Cvc only form with field type = CREDIT_CARD_VERIFICATION_CODE
(this is when VCN CoF flag is disabled)
    FormEvents logged to "Autofill.FormEvents.CreditCard" (It should
    be StandaloneCvc, I will update in following CL)
3) Cvc only form with field type = CREDIT_CARD_STANDALONE_VERIFICATION_CODE
    FormEvents logged to "Autofill.FormEvents.StandaloneCvc"

This is a preparation CL for the new logging logic for Cvc only form.

dd: go/pay-autofill-cvc-storage-enhancement-dd

Bug: 356694842
Change-Id: Ifb5ea31dd7d377f9ca45ebf2fdc3adc62a28f662
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5824089
Reviewed-by: Jan Keitel <jkeitel@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Alexander Tekle <alexandertekle@google.com>
Reviewed-by: Olivia Saul <jsaul@google.com>
Commit-Queue: Sujie Zhu <sujiezhu@google.com>
Cr-Commit-Position: refs/heads/main@{#1352330}